Add a test configuration for the API to docker-compose
This seems to be slightly less confusing than having the API start up
two different Elasticsearch containers and (hopefully) only direct
traffic to one of them.

In the tests, we were apparently sending traffic to both and the results
were very hard to debug. With this change we will use the same port
number for test and production Elasticsearch hosts, but we will use a
different host name for each.